refactor: move usage json pricing into the library
Add usage::price_usage_data returning a serializable PricedUsageRow so a
non-CLI consumer can build the usage --json payload, and derive Serialize
on the top-level usage/analysis summary types. The binary keeps the same
output shape.
